// src/infra/http/services/api/modules/types/gateway-theme.types.ts
export interface IGatewayThemeColors {
  id: number;
  primaryColor: string;
  primaryForeground: string;
  backgroundColor: string;
  foregroundColor: string;
  cardColor: string;
  cardForeground: string;
  borderColor: string;
  mutedColor: string;
  mutedForeground: string;
  accentColor: string;
  accentForeground: string;
  destructiveColor: string;
  successColor: string;
  warningColor: string;
  darkPrimaryColor: string;
  darkBackgroundColor: string;
  darkForegroundColor: string;
  darkCardColor: string;
  darkCardForeground: string;
  darkBorderColor: string;
  darkMutedColor: string;
  darkMutedForeground: string;
  darkAccentColor: string;
  darkAccentForeground: string;
}